requirements.txt
beautifulsoup4
GitPython
invoke
lxml
more-itertools
pymediawiki
